Add files via upload
GameDynamicsForTest: The game dynamics needed to be updated to run the AI vs AI (This is not needed in S or M mode)
AIPlayer: Updated, should have level one and not break down
AIPlayerForTest: This is the exact same as AIPlayer but has a bunch of print statements that illustrate what is going on 
TestAIVsAI: Uses the above mentioned test classes to run iterations of playing the game. Try it out!